backend/requirements.txt
aiosqlite==0.19.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
extruct==0.16.0
fake-useragent==1.2.1
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
html-text==0.5.2
html5lib==1.1
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
jq==1.6.0
jstyleson==0.0.2
lxml==4.9.3
mf2py==1.1.3
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyee==8.2.2
PyJWT==2.8.0
pyparsing==3.1.1
pyppeteer==1.0.2
pyRdfa3==3.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
rdflib==7.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==1.26.16
uvicorn==0.23.2
w3lib==2.1.2
webencodings==0.5.1
websockets==10.4
zipp==3.17.0
